+++ src/com/magento/idea/magento2plugin/MagentoIcons.java
@@ -1,3 +1,7 @@
+/**
+ * Copyright © Dmytro Kvashnin. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 package com.magento.idea.magento2plugin;
 
 import com.intellij.openapi.util.IconLoader;
@@ -9,4 +13,5 @@ import javax.swing.*;
  */
 public class MagentoIcons {
     public static final Icon WEB_API = IconLoader.getIcon("icons/webapi.png");
+    public static final Icon MODULE = IconLoader.getIcon("/icons/module.png");
 }