+++ src/com/magento/idea/magento2plugin/indexes/IndexManager.java
@@ -1,8 +1,16 @@
+/**
+ * Copyright © Dmytro Kvashnin. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 package com.magento.idea.magento2plugin.indexes;
 
 import com.intellij.util.indexing.FileBasedIndexImpl;
 import com.intellij.util.indexing.ID;
 import com.magento.idea.magento2plugin.stubs.indexes.*;
+import com.magento.idea.magento2plugin.stubs.indexes.js.MagentoLibJsIndex;
+import com.magento.idea.magento2plugin.stubs.indexes.js.RequireJsIndex;
+import com.magento.idea.magento2plugin.stubs.indexes.graphql.GraphQlResolverIndex;
+import com.magento.idea.magento2plugin.stubs.indexes.mftf.*;
 import com.magento.idea.magento2plugin.stubs.indexes.xml.PhpClassNameIndex;
 
 /**
@@ -25,7 +33,18 @@ public class IndexManager {
             // webapi
             WebApiTypeIndex.KEY,
             ModuleNameIndex.KEY,
-            PhpClassNameIndex.KEY
+            PhpClassNameIndex.KEY,
+            //require_js
+            RequireJsIndex.KEY,
+            MagentoLibJsIndex.KEY,
+            // mftf
+            ActionGroupIndex.KEY,
+            DataIndex.KEY,
+            PageIndex.KEY,
+            SectionIndex.KEY,
+            StepKeyIndex.KEY,
+             //graphql
+            GraphQlResolverIndex.KEY
         };
 
         for (ID<?, ?> id: indexIds) {