+++ src/com/magento/idea/magento2plugin/php/module/ComposerPackageModel.java
@@ -1,3 +1,7 @@
+/**
+ * Copyright © Dmytro Kvashnin. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 package com.magento.idea.magento2plugin.php.module;
 
 import org.jetbrains.annotations.Nullable;