+++ src/com/magento/idea/magento2plugin/stubs/indexes/ModuleNameIndex.java
@@ -1,3 +1,7 @@
+/**
+ * Copyright © Dmytro Kvashnin. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 package com.magento.idea.magento2plugin.stubs.indexes;
 
 import com.intellij.openapi.util.text.StringUtil;